[sprint_5-game_start_screen] вынос констант в отдельную папку
