+++ packages/client/src/constants/constants.tsx
@@ -1,0 +1,8 @@
+export const GAME_DESCRIPTION = {
+    name: 'Destroy This Bubbles',
+    rules: [
+        'Щёлкните по пузырям, чтобы уничтожить их',
+        'Чем больше пузырей уничтожаете, тем больше очков вы получаете',
+        'За пропущенные пузыри вы теряете очки',
+    ],
+};
