+++ packages/client/src/pages/game/launcher/launcher.tsx
@@ -1,5 +1,4 @@
-import { GAME_DESCRIPTION } from 'constants';
-
+import { GAME_DESCRIPTION } from '@constants/constants';
 import { Button } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
